Fix import dialog race conditions and prevent duplicate callbacks
- Add guard to prevent calling onSuccess/onOpenChange twice
- Fix close button state when character wizard is open
- Add error handling for character detection failures
- Move character detection types to shared package
- Use unique IDs for checkboxes to prevent collisions
