+++ apps/frontend/src/lib/api/characters.ts
@@ -5,36 +5,15 @@
  */
 
 import { request } from "./client";
-import type { Character } from "@branchforge/shared";
+import type {
+  Character,
+  DetectedCharacter,
+  CharacterConflict,
+  DetectCharactersResponse,
+} from "@branchforge/shared";
 
-// ============================================================================
-// Types
-// ============================================================================
-
-export interface DetectedCharacter {
-  tag: string;
-  name: string | null;
-  displayName: string;
-  color: string;
-  isSpecial: boolean;
-  sourceFile: string;
-  confidence: number;
-}
-
-export interface CharacterConflict {
-  tag: string;
-  detectedName: string | null;
-  existingName: string;
-  detectedColor: string;
-  existingColor: string;
-}
-
-export interface DetectCharactersResponse {
-  characters: DetectedCharacter[];
-  excludedTags: string[];
-  conflicts: CharacterConflict[];
-  existingTags: string[]; // Tags of all characters that exist in database (not just conflicts)
-}
+// Re-export types for convenience
+export type { DetectedCharacter, CharacterConflict, DetectCharactersResponse };
 
 export interface ImportCharacter {
   tag: string;
